requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.6
executing==1.2.0
idna==3.4
inflection==0.5.1
ipython==8.14.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
openapi-codec==1.3.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.6
bcrypt
djangorestframework-simplejwt
